Docker_Kubernetes_manifests/Backend_File_Structure/requirements.txt
fastapi
uvicorn
pydantic
requests
PyJWT~=2.4.0
cryptography
pyyaml
aiohttp
asyncio